Do not derive MIME type from image reference URL
I discovered that UIImagePickerControllerReferenceURL was not a good source to derive the picked media item's MIME type from, because UIImagePickerControllerReferenceURL will be nil for for newly captured images. I also discovered that react-native-image-picker always creates either a PNG or JPEG representation of picked still images, so I decided to just set the MIME types on the response object for these data presentations when they are created.